requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
oauthlib==3.2.2
py-trello==0.19.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
urllib3==2.0.7
